+++ backend/ntags/models.py
@@ -6,80 +6,8 @@ from django.utils.translation import gettext_lazy as _
 
 from .constants import NTAG213, NTAG_IC_CHOICES, NTAG_EEPROM_SIZES
 from .validators import validate_serial_number, validate_integrated_circuit
-from .utils import get_nfc_tag_model_string, get_nfc_tag_filter_method
-
-
-class NFCTagManager(models.Manager):
-    def __init__(self, through=None, model=None, instance=None):
-        self.through = through
-        self.model = model
-        self.instance = instance
-        super().__init__()
-
-    def __get__(self, instance, model):
-        manager = NFCTagManager(
-            through=self.through,
-            model=self.model,
-            instance=instance
-        )
-        return manager
-
-    def get_from_mirror(self, ascii_mirror):
-        """
-        Retrieves an NFCTag instance based on the mirrored ASCII value and filter type.
-        """
-        if not ascii_mirror:
-            raise ValueError(_('Invalid NFC tag URI.'))
-
-        ntag_filter = get_nfc_tag_filter_method()
-
-        if ntag_filter == 'uid':
-            return self.get_from_uid(ascii_mirror)
-        elif ntag_filter == 'counter':
-            return self.get_from_counter(ascii_mirror)
-        elif ntag_filter == 'uid_counter':
-            return self.get_from_uid_counter(ascii_mirror)
-        else:
-            raise ValueError(_('Invalid filter method.'))
-
-    def get_from_uid(self, uid):
-        """
-        Retrieves an NFCTag instance based on the UID.
-        """
-        try:
-            return self.get(serial_number=uid)
-        except self.model.DoesNotExist:
-            raise self.model.DoesNotExist(_('NFC Tag not found.'))
-
-    def get_from_counter(self, counter):
-        """
-        Retrieves an NFCTag instance based on the Counter.
-        """
-        from .models import NFCTagScan
-        latest_scan_subquery = NFCTagScan.objects.filter(
-            ntag=models.OuterRef('pk')
-        ).order_by('-scanned_at').values('counter')[:1]
-
-        # Annotate NFCTag with the latest counter value
-        queryset = self.annotate(
-            latest_counter=models.Subquery(latest_scan_subquery)
-        ).filter(
-            latest_counter=counter
-        )
-        return queryset
-
-    def get_from_uid_counter(self, uid_counter):
-        """
-        Retrieves an NFCTag instance based on the UID and counter.
-        """
-        if 'x' not in uid_counter:
-            raise ValueError(_('Invalid NFC Tag Mirror.'))
-        uid, counter = uid_counter.split('x')
-
-        try:
-            return self.get(serial_number=uid).scans.get(counter=counter)
-        except self.model.DoesNotExist:
-            raise self.model.DoesNotExist(_('NFC Tag Scan not found.'))
+from .utils import get_nfc_tag_model_string
+from .managers import NFCTagManager
 
 
 class AbstractNFCTag(models.Model):
@@ -108,6 +36,8 @@ class AbstractNFCTag(models.Model):
         related_name='ntags'
     )
 
+    objects = NFCTagManager()
+
     def log_scan(self, counter):
         from .models import NFCTagScan
         return NFCTagScan.objects.create(
@@ -125,6 +55,9 @@ class AbstractNFCTag(models.Model):
 
 
 class NFCTag(AbstractNFCTag):
+    """
+    Concrete model for NFC Tags.
+    """
     pass
 
 
@@ -172,7 +105,6 @@ class NFCTagMemory(models.Model):
         return str(self.ntag)
 
     class Meta:
-        abstract = True
         verbose_name = _("eeprom")
         verbose_name_plural = _("eeproms")
 
@@ -204,6 +136,5 @@ class NFCTagScan(models.Model):
         return f"Scan #{self.counter} for {self.ntag}"
 
     class Meta:
-        abstract = True
         verbose_name = _("scan")
         verbose_name_plural = _("scans")
