+++ backend/ntags/views.py
@@ -4,7 +4,6 @@ from django.utils.translation import gettext_lazy as _
 
 from .utils import get_nfc_tag_model
 
-
 def link(request):
     """
     Handles the initial logic for an NFC tag link.
